refactor: change structure to modular structure, mvc as pattern, apply some best practices
